requirements.txt
azure_storage==0.37.0
msrest==0.7.0
Pillow==9.1.1
requests==2.28.0
